Revert "Feat: added docs framework (#15)"
This reverts commit 0e4a92b344d07f7521249f0738407ee3494be4bc.
