+++ .editorconfig
@@ -3,6 +3,6 @@ charset = utf-8
 insert_final_newline = true
 end_of_line = lf
 
-[*.{json,js,ts,tsx,rs,codegen}]
+[*.{json,js,ts,tsx,rs}]
 indent_size = 4
 indent_style = tab
