+++ .gitignore
@@ -51,6 +51,3 @@ yalc.lock
 
 #ignore output directory
 /build
-
-#ignore deploy config
-deploy.json
