+++ backend/build.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+cargo build --release
+mkdir -p ../bin
+
+cp ./target/release/backend ../bin/backend
