--- deploy.example.json
@@ -1,5 +1,0 @@
-{
-	"host": "steamdeck",
-	"user": "deck",
-	"keyfile": "~/.ssh/id_rsa"
-}
