--- docs/Credits.mdx
@@ -1,9 +1,0 @@
-# Credits
-
-Massive shout outs to:
-
-- [@jeiza]()
-- [@rainbyte]()
-- [@nabelo]()
-
-without whom this plugin just simply wouldn't exist as it does today.
