--- docs/index.mdx
@@ -1,3 +1,0 @@
-# Hello, World!
-
-This is the beginning of the docs, Unfortunately they are still under construction ⚠️🚧👷
