+++ src/index.tsx
@@ -15,7 +15,7 @@ import PatchAppScreen from "./patch/PatchAppScreen";
 import { API_URL, DOCUMENTATION_PATH, UNAMED_CARD_NAME } from "./const";
 import { Logger } from "./Logging";
 import React from "react";
-import Docs from "./pages/Docs";
+import DocumentationPage from "./pages/Docs";
 import { DeckyAPI } from "./lib/DeckyApi";
 import { MicroSDeck, MicroSDeckContextProvider, useMicroSDeckContext, CardAndGames, MicroSDCard, IsMatchingSemver } from "../lib/src";
 import { CardActionsContextMenu } from "./components/CardActions";
@@ -141,6 +141,9 @@ declare global {
 }
 
 export default definePlugin((serverApi: ServerAPI) => {
+	serverApi.routerHook.addRoute(DOCUMENTATION_PATH, DocumentationPage, {
+		exact: true,
+	});
 
 	if (window.MicroSDeck) {
 		window.MicroSDeck.destruct();
@@ -149,11 +152,9 @@ export default definePlugin((serverApi: ServerAPI) => {
 
 	DeckyAPI.SetApi(serverApi);
 
-	Logger.Log("Started MicroSDeck");
-	
 	const patch = PatchAppScreen(serverApi);
 
-	serverApi.routerHook.addRoute(DOCUMENTATION_PATH, Docs);
+	Logger.Log("Started MicroSDeck");
 
 	return {
 		title: <div className={staticClasses.Title}>MicroSDeck</div>,
