+++ tsconfig.json
@@ -1,25 +1,25 @@
 {
-	"compilerOptions": {
-		"outDir": "build/dist",
-		"module": "ESNext",
-		"target": "ES2020",
-		"jsx": "react",
-		"jsxFactory": "window.SP_REACT.createElement",
-		"jsxFragmentFactory": "React.Fragment",
-		"declaration": false,
-		"moduleResolution": "nodenext",
-		"resolveJsonModule": true,
-		"noUnusedLocals": true,
-		"noUnusedParameters": true,
-		"esModuleInterop": true,
-		"noImplicitReturns": true,
-		"noImplicitThis": true,
-		"noImplicitAny": true,
-		"strict": true,
-		"suppressImplicitAnyIndexErrors": true,
-		"allowSyntheticDefaultImports": true,
-		"skipLibCheck": true
-	},
-	"include": ["src", "lib"],
-	"exclude": ["node_modules"]
+  "compilerOptions": {
+    "outDir": "build/dist",
+    "module": "ESNext",
+    "target": "ES2020",
+    "jsx": "react",
+    "jsxFactory": "window.SP_REACT.createElement",
+    "jsxFragmentFactory": "React.Fragment",
+    "declaration": false,
+    "moduleResolution": "nodenext",
+    "resolveJsonModule": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "esModuleInterop": true,
+    "noImplicitReturns": true,
+    "noImplicitThis": true,
+    "noImplicitAny": true,
+    "strict": true,
+    "suppressImplicitAnyIndexErrors": true,
+    "allowSyntheticDefaultImports": true,
+    "skipLibCheck": true
+  },
+  "include": ["src", "lib"],
+  "exclude": ["node_modules"]
 }
