--- util/log.mjs
@@ -1,10 +1,0 @@
-import Lipe, { LoggerPipe } from "lipe"
-import { PrefixWithColor, Splat, Timestamped, Console } from "lipe/defaults"
-
-const Pipe = new LoggerPipe()
-	.Pipe(PrefixWithColor)
-	.Pipe(Splat("{prefix} {Message}"))
-	.Pipe(Console())
-
-
-export const Logger = new Lipe().AddPipe(Pipe);
